+++ Source/ReleaseNotes.md
@@ -2,6 +2,7 @@
 * Core: Performance improvements
 * Core: Added validations for variable byte integers which do not match the .NET uint perfectly
 * Client: Added support for pre-encoded UTF-8 binary buffers for user properties (#2228, thanks to @koepalex)
+* Client: Late acknowledgement packets (PUBACK, PUBCOMP, SUBACK, UNSUBACK) for cancelled or timed-out requests no longer trigger a protocol violation and tear down the connection (#2078, #2079)
 * Server: Improved performance of retained messages when no event handler is attached (#2093, thanks to @zhaowgit)
 * Server: The event `InterceptingClientEnqueue` is now also called for retained messages (BREAKING CHANGE!)
 * Server: The local end point is now also exposed in the channel adapter (#2179)